further splitting of the pipeline

+++ src/main.rs
@@ -110,7 +110,7 @@ fn parse_cmd(
             for (idx, str) in rd.words.iter().enumerate().filter(|v| v.1.contains(&fval)) {
                 println!("    {} {}", str, rd.word_count[idx]);
                 for f_idx in &rd.file_idx[idx] {
-                    println!("         {}", rd.files[*f_idx]);
+                    println!("         {}", rd.files[*f_idx as usize]);
                 }
             }
         }
